DTSAM-1006 Centralise the common code for Process 6.
